0.19.11 new package for Talents + refactoring
