+++ wfrp/src/main/java/com/warhammer/wfrp/engine/TalentsGenerator.java
@@ -1,11 +1,10 @@
 package com.warhammer.wfrp.engine;
 
-import com.warhammer.wfrp.service.TalentService;
+import com.warhammer.wfrp.talents.TalentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor
@@ -33,7 +32,7 @@ public class TalentsGenerator {
                 return generateWoodElfTalents();
             }
             default -> {
-                return null;
+                return Collections.emptyList();
             }
         }
     }
