+++ wfrp/src/main/java/com/warhammer/wfrp/talents/TalentRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.talent;
+package com.warhammer.wfrp.talents;
 
-import com.warhammer.wfrp.model.talent.Talent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TalentRepository extends JpaRepository<Talent, Integer> {
